refactor!: refactor code structure
The goal is to keep related code together, instead of spreading it out into
top-level modules exclusively.

This also renames `Index` to `GitIndex`.
Further changes involved renaming `ChangesIter` to `git::Changes`, and
`INDEX_GIT_URL` to `git::URL`, and `CRATES_IO_HTTP_INDEX` to `sparse::URL`.
